% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_oldest.R
\name{find_oldest}
\alias{find_oldest}
\title{A Function to Find Oldest NBA Player by Year}
\usage{
find_oldest(yearinput)
}
\description{
This function allows you to Find Oldest NBA Player by Year
}
\examples{
find_oldest()
}
\keyword{nba,oldest}
